+++ app/app.js
@@ -9,6 +9,7 @@ window.run = function() {
         'config/baseConfig',
         'config/headerConfig',
         'plugins/anchoredLayoutPlugin',
+        'plugins/mobifyPreviewPlugin',
         'controllers/counterBadgeController',
         'app-components/deepLinkingServices',
         'app-controllers/tabBarController',
@@ -25,6 +26,7 @@ window.run = function() {
         BaseConfig,
         HeaderConfig,
         AnchoredLayoutPlugin,
+        MobifyPreviewPlugin,
         CounterBadgeController,
         DeepLinkingServices,
         TabBarController,
@@ -123,7 +125,10 @@ window.run = function() {
             });
         };
 
-        var runApp = function() {
+        var runApp = function(previewedUrl) {
+            // TODO: [HYB-884] As a Scaffold developer,
+            // I would like for the baseURL to be set to the previewed URL
+
             WelcomeModalController.init(errorControllerPromise)
                 .then(function(welcomeModalController) {
                     // The welcome modal can be configured to show only once
@@ -135,7 +140,19 @@ window.run = function() {
                 });
         };
 
-        runApp();
+        var runAppPreview = function() {
+            MobifyPreviewPlugin.init()
+                .then(function(previewPlugin) {
+                    previewPlugin
+                        .preview(BaseConfig.baseURL, BaseConfig.previewBundle)
+                        .then(runApp);
+                });
+        };
+
+        // Configure the previewEnabled flag located in baseConfig.js
+        // to enable/disable app preview
+        BaseConfig.previewEnabled ? runAppPreview() : runApp();
+
     }, undefined, true);
 };
 // Comment out next line for JS debugging