+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## To be released
+- Integrate the Mobify preview plugin
 - Welcome modal now relies on installationId to determine whether to show
 - Add Requirements to README, and "engines" field to package.json